% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{extract.cols}
\alias{extract.cols}
\title{Extract columns based on param name}
\usage{
extract.cols(dt, param)
}
\arguments{
\item{dt}{The data.table from where to extract the columns}

\item{param}{The name of the param used internally}
}
\description{
Extract columns based on param name
}
