% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{extract.HPD}
\alias{extract.HPD}
\title{Extract the Highest Probability Density (HPD) from logged data}
\usage{
extract.HPD(log, alpha = 0.05)
}
\arguments{
\item{log}{The log (read with read.log)}

\item{alpha}{The credibility interval (between 0 and 1)}
}
\value{
A data.table with HPD (lower and upper limits, and median)
}
\description{
This function uses boa::boa.hpd
}
