% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{extract.param}
\alias{extract.param}
\title{Extract a parameter from HPD using a regular expression}
\usage{
extract.param(HPD, param)
}
\arguments{
\item{HPD}{The HPD table (obtained with extract.HPD)}

\item{param}{The name (pattern) to use in the regular expression}
}
\value{
A data.table
}
\description{
Extract a parameter from HPD using a regular expression
}
