% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{make.BDMMbyDeme.skyplot}
\alias{make.BDMMbyDeme.skyplot}
\title{Produce a skyline plot starting from a BDMM-prime MCMC log}
\usage{
make.BDMMbyDeme.skyplot(
  log,
  par = c("R0", "s", "delta", "r"),
  nBins = 100,
  type = c("smooth", "step"),
  alpha = 0.05,
  maxAge
)
}
\arguments{
\item{log}{The log (read with read.log)}

\item{par}{the parameter to be plotted. Possible option: "R0", "s" (sampling
proportion), "delta" (becomingUninfectiousRate), "r" (removevalProbability)}

\item{nBins}{the number of bins to use (integer of length=1)}

\item{type}{The type of plot to generate. This will depend on how the
analysis has been set up: "smooth" if the timeChanges parameters are
estimated; "step" is these are fixed.}

\item{alpha}{The credibility interval (between 0 and 1)}

\item{maxAge}{The age of the most recent sample.}
}
\description{
This function deals with parameters that change over time within each deme
(e.g. R0, s), not between demes (e.g. migration R0AmongDemes).
}
\details{
It returns a ggplot with overimposed (with transparency) plot colour-coded
for each deme. This can be edited post-production as usual for
ggplot-objects. For example, if multiple plots - one for each deme - are
needed, this can be easily achieved with \code{facet_grid} (say the plot generated by \code{make.BDMMbydeme.skyplot} is p, you could use :
\code{p + facet_grid(Deme~.)} to have each deme on separate rows).
}
