% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.BDMMbyDeme.violin.R
\name{make.BDMMbyDeme.violin}
\alias{make.BDMMbyDeme.violin}
\title{Draw a violin plot for each deme and epoch}
\usage{
make.BDMMbyDeme.violin(log, par = c("R0", "s", "delta", "r"), keepScale = TRUE)
}
\arguments{
\item{log}{The log (read with read.log)}

\item{par}{the parameter to be plotted. Possible option: "R0", "s" (sampling
proportion), "delta" (becomingUninfectiousRate), "r" (removevalProbability)}

\item{keepScale}{Whether the y axis should be kept consistent (TRUE) across
all rows of the plot or adjusted depending on the posterior distribution
(FALSE). The second is sometime more convenient for readability.}
}
\value{
a ggplot
}
\description{
This function generates a violin plot with each deme is plotted (and
colour-coded) separately. Each Epoch is plotted as one line in the plot.
}
\details{
Within the violin plot there are tree lines, the two most external lines mark
the 2.5 and 97.5 percentile. The middle one is the median.

The object returned is a ggplot object, which can be modified as usual. For
example, in presence of long tails, one may want to trim the y axis like so:
\code{p + ylim(c(l, u))}, where p is the plot returned by the function, and
\code{l} and \code{u} are the lower and upper desire limits on the y axis. It
is important to also note that the first element of the ggplot list returned
are the actual data that are used to generate the plot, which can then be
used for other down-stream analysis or plots.
}
