% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{make.epochs.data}
\alias{make.epochs.data}
\title{Generate the starting and ending times of the epochs}
\usage{
make.epochs.data(HPD, maxAge, gridPoints, rootHeight, reverse = FALSE)
}
\arguments{
\item{maxAge}{The age of the most recent sample.}

\item{gridPoints}{The times used in the analysis as gridPoints (skygrid) or
changeTimes (BD models)}

\item{rootHeight}{The median value of the rootHeight}

\item{reverse}{When \code{FALSE}, the epoch in the log are from the oldest to
the youngest (that is, the first entry for the parameter is for the oldest
epoch). If \code{TRUE} the epoch in the log are from the youngest to the
oldest.}
}
\value{
A data.table
}
\description{
skygrid reports as first parameter the most recent, so needs \code{reverse=TRUE}.
}
