% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{make.skyplot}
\alias{make.skyplot}
\title{Generate a skyline plot}
\usage{
make.skyplot(epochData, by = NULL)
}
\arguments{
\item{epochData}{The starting and ending times of the epochs (generated with
make.epochs.data) or a list (see Details)}

\item{by}{the column's name to use to colourcode the skylines}
}
\description{
If more than one set of data are passed, the function will overimpose each
set in the same plot. In this case, the data have to be in a list, with each
data.table as an element of the list, and an additional column needs to be
added to the data. This column, whose name has to be passed with the argument
\code{by}, will be used to colourcode the plot and identify the datasets.
}
\details{
The plot returned by the function is a ggplot object and can, therefore, be
modified as usual (e.g. \code{plot + theme_classic()})
}
