% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDSkyGridScripts.R
\name{read.log}
\alias{read.log}
\title{Reads a log file}
\usage{
read.log(filename, burnin = 0.1)
}
\arguments{
\item{filename}{The name of the file (with the path if necessary)}

\item{burnin}{The burnin to remove from the log in proportion}
}
\value{
A data.table
}
\description{
Reads a log file
}
